Adding the dist/ so the user of this project won't have to manually create it.
The project where I'm using this is not auto-creating the dist/ when
calling `npm install` in that parent project. I can cd into the
node_modules/swagger-ui-browserify/ and manually call `npm install` to
get the dist/, but this will be a pain for other developers and
automating the deployment of the project. I was unable to figure out how
to get the dist/ to be created by the parent project, so I'm adding it
to version control here in order to get it to be immediately available
in the parent project.